Implimentation
First of all, I implimented a hover photo transition with the help of w3schools and some other links that will be placed as refererce within the readme.md. Also deleted the nav.js due to not doing what i want and i wasnt able to find it on the internet. i wanted to set the navigation on every page with a placeholder like in rubyonrails but couldnt get it to function properly. Other changes are with css codes and implimenting color changes and so fort.