Filtering Page
Still trying to style the Anime Page and to also impliment it to how i want it to look. At the moment i'm changing the images to the middle within the boxes. With the titles on top. Also trying to impliment arrows within the boxes but having difficulty getting the images to show from using url(...images) from the style sheet.