Problems (Restructuring)
Still been trying to impliment a filter function that doesnt use the w3school onclick. But been having problems restructuring it and change it to match how i want the coding to work. I give up. So i'll either have to start the other pages and come back to the filtering system or leave it entirely. :(