Javascript & Html
For some reason i can not get the filtering js code to work from a external source. I'm sure that this might be a simple fix because i resolved it before but for some reason it chooses not to work today. So will try again later. As for now i'll keep it running within the html.