Removed Js from HTML
Before this wasnt working but now somehow it is. Probably was a problem with browsers. Not really sure or i was up too late and tired and coundlt see the error. Anyways it works now without it being in the html file. Now after i finish everything i can figure out how to remove the onclicks.