Googlemap
I've seen once where to put in a googlemap one needed some js code for it to function. But i've found this way which seems more efficent and works properly. it only had given my some instyle along with the code but i figured out how to style it from the css file easily.