Nu Html Checker
Finished Running and correcting the html checker. In Total the html pages only the index page is left with a "Section lacks heading. Consider using h2-h6 elements to add identifying headings to all sections." But I've also researched where it is said that within HTML 5 this error can be overlooked. Also i need the section where it is.